Backed out changeset 9093e927a247 (bug 1201997)
UltraBlame original commit: 816b6cd56bd245e2f497b7a81113032d7b87a2a6

+++ netwerk/sctp/datachannel/DataChannel.cpp
@@ -1849,7 +1849,7 @@ DataChannelConnection::HandleStreamChangeEvent(const struct sctp_stream_change_e
 
       
       
-      size_t num_needed = mPending.GetSize();
+      int32_t num_needed = mPending.GetSize();
       LOG(("%d of %d new streams already needed", num_needed,
            new_len - old_len));
       num_needed -= (new_len - old_len);
